Completed Authentication functionality

+++ controller/expense.js
@@ -14,20 +14,22 @@ const addexpense = (req,res)=>{
         return res.status(400).json({err:'Fields left empty'})
     }
 
-    Expense.create({expenseamount,description,category}).then(expense =>{
+    Expense.create({expenseamount,description,category,userId:req.user.id}).then(expense =>{
         return res.status(201).json({expense, success:true})
     })
     .catch(err=>{
+        console.log(err)
         return res.status(500).json({success :false,error:err})
     })
 }
 
 const getexpenses = (req,res)=>{
 
-    Expense.findAll().then(expenses=>{
+    Expense.findAll({where :{userId:req.user.id}}).then(expenses=>{
         return res.status(200).json({expenses,success:true})
     })
     .catch(err=>{
+        console.log(err)
         return res.status(500).json({error:err,success:false})
     })
 }
@@ -37,7 +39,7 @@ const deleteexpense = (req,res)=>{
  if(stringvalidator(expenseid)){
   return  res.status(400).json({success:false,message:"Id not found"})
  }
- Expense.destroy({where:{id:expenseid}})
+ Expense.destroy({where:{id:expenseid,userId:req.user.id}})
  .then(()=>{
     return res.status(200).json({success:true,message:"Delete Successful"})
  })
